Move `DiscoveredSymbol[s]` into `Types`
