+++ src/Console/Commands/Compose.php
@@ -10,12 +10,12 @@ use BrianHenryIE\Strauss\Composer\Extra\StraussConfigInterface;
 use BrianHenryIE\Strauss\Composer\ProjectComposerPackage;
 use BrianHenryIE\Strauss\Copier;
 use BrianHenryIE\Strauss\DependenciesEnumerator;
-use BrianHenryIE\Strauss\DiscoveredSymbols;
 use BrianHenryIE\Strauss\FileEnumerator;
 use BrianHenryIE\Strauss\Files\DiscoveredFiles;
 use BrianHenryIE\Strauss\FileScanner;
 use BrianHenryIE\Strauss\Licenser;
 use BrianHenryIE\Strauss\Prefixer;
+use BrianHenryIE\Strauss\Types\DiscoveredSymbols;
 use Exception;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LogLevel;
