+++ src/Types/ConstantSymbol.php
@@ -2,8 +2,6 @@
 
 namespace BrianHenryIE\Strauss\Types;
 
-use BrianHenryIE\Strauss\DiscoveredSymbol;
-
 class ConstantSymbol extends DiscoveredSymbol
 {
